Updated user profile photo upload
